Menambahkan api route serta unit testing-nya

+++ app.js
@@ -10,25 +10,27 @@ const middlewaresError = require('./middlewares/error');
 
 const app = express();
 
-// DB connection
-(async function () {
-  try {
+if(process.env.NODE_ENV !== 'test') {
+  // DB connection
+  (async function () {
+    try {
 
-    await db.sequelize.sync({ force: true });
+      await db.sequelize.sync({ force: true });
 
-    // await db.sequelize.sync();
-    await db.sequelize.authenticate();
-    console.log("Connection has been established successfully.");
-    
-    const user = await db.User.create({
-        username: 'user test',
-        password: 'slslwww'
-    });
-    
-  } catch (error) {
-    console.error(error);
-  }
-})();
+      // await db.sequelize.sync();
+      await db.sequelize.authenticate();
+      console.log("Connection has been established successfully.");
+      
+      const user = await db.User.create({
+          username: 'user test',
+          password: 'slslwww'
+      });
+      
+    } catch (error) {
+      console.error(error);
+    }
+  })();
+}
 
 // Middleware
 app.use(morgan('dev'));
